read control packet feature

+++ NPUproject/twoCTPassThrough/host/typedef.hpp
@@ -17,8 +17,8 @@
 #include "xrt/xrt_kernel.h"
 #include "buffer.hpp"
 
-typedef float dtype_in;
-typedef float dtype_out;
+typedef uint32_t dtype_in;
+typedef uint32_t dtype_out;
 
 
 #endif
